Add 'leave' status & per-session attendance stats
Introduce a 'leave' status across the UI and data model, compute absences per-session (ensuring non-negative counts), and surface per-session statistics in ClassDetailView (latest session overview, stacked 5-session chart, limited session selector, and various UI/UX tweaks). Update StudentDashboard to fetch sessions, compute attended counts/total sessions, expose latestSession/latestRecord, and calculate attendance percentage based on total sessions. Also change FastAPI/Pydantic duration_hours from int to float in main_refactored.py (and corresponding form params) to support fractional hour durations.

+++ face_recognition_server/main_refactored.py
@@ -91,12 +91,12 @@ class StartStreamRequest(BaseModel):
     class_id: str
     teacher_email: str
     on_time_limit_minutes: int = 15
-    duration_hours: int = 2
+    duration_hours: float = 2
 
 class MotionSessionRequest(BaseModel):
     class_id: str
     teacher_email: str
-    duration_hours: int = 2
+    duration_hours: float = 2
     motion_threshold: float = 0.1
     cooldown_seconds: int = 30
     on_time_limit_minutes: int = 30
@@ -184,7 +184,7 @@ async def root():
 async def start_motion_detection(
     class_id: str = Form(...),
     teacher_email: str = Form(...),
-    duration_hours: int = Form(2),
+    duration_hours: float = Form(2),
     motion_threshold: float = Form(0.1),
     cooldown_seconds: int = Form(30),
     on_time_limit_minutes: int = Form(30)
@@ -776,7 +776,7 @@ async def start_realtime_stream(
     class_id: str = Form(...),
     teacher_email: str = Form(...),
     on_time_limit_minutes: int = Form(15),
-    duration_hours: int = Form(2)
+    duration_hours: float = Form(2)
 ):
     """Alias for start_motion_session to match TypeScript interface"""
     return await start_motion_session(
